gov_QA_portal/requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
argcomplete==3.1.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Bottleneck==1.3.5
Brlapi==0.8.5
Brotli==1.1.0
cachetools==5.3.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
et-xmlfile==1.0.1
fasteners==0.18
fonttools==4.46.0
frozenlist==1.4.0
fs==2.4.16
google-api-python-client==1.7.12
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
greenlet==3.0.3
gspread==6.2.1
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.3.1
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==0.0.0
multidict==6.0.4
mutagen==1.46.0
numexpr==2.9.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4+dfsg
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.4
proton-vpn-gtk-app==4.9.6
proton-vpn-network-manager==0.12.13
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pycares==4.4.0
pycryptodomex==3.20.0
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-gnupg==0.5.2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
SciPy==1.11.4
SecretStorage==3.3.3
sentry-sdk==1.39.2
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.50
sympy==1.12
systemd-python==235
tables==3.9.2
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
uritemplate==4.1.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.8.2
yt-dlp==2024.4.9
zipp==1.0.0
